+++ plugins/scaffolder/src/alpha/apis.tsx
@@ -19,10 +19,69 @@ import {
   discoveryApiRef,
   fetchApiRef,
   identityApiRef,
+  appTreeApiRef,
+  createExtensionInput,
 } from '@backstage/frontend-plugin-api';
 import { scmIntegrationsApiRef } from '@backstage/integration-react';
 import { scaffolderApiRef } from '@backstage/plugin-scaffolder-react';
+import {
+  FormFieldBlueprint,
+  formFieldsApiRef,
+} from '@backstage/plugin-scaffolder-react/alpha';
 import { ScaffolderClient } from '../api';
+import { OpaqueFormField } from '@internal/scaffolder';
+
+export const formFieldsApi = ApiBlueprint.makeWithOverrides({
+  name: 'form-fields',
+  inputs: {
+    formFields: createExtensionInput([
+      FormFieldBlueprint.dataRefs.formFieldLoader,
+    ]),
+  },
+  factory(originalFactory, { inputs }) {
+    const formFieldLoaders = inputs.formFields.map(e =>
+      e.get(FormFieldBlueprint.dataRefs.formFieldLoader),
+    );
+
+    return originalFactory(defineParams =>
+      defineParams({
+        api: formFieldsApiRef,
+        deps: { appTreeApi: appTreeApiRef },
+        factory: ({ appTreeApi }) => ({
+          async loadFormFields() {
+            const pageFormFieldLoaders = getPageFormFieldLoaders(appTreeApi);
+
+            const formFields = await Promise.all(
+              [...formFieldLoaders, ...pageFormFieldLoaders].map(loader =>
+                loader(),
+              ),
+            );
+
+            return formFields.map(OpaqueFormField.toInternal);
+          },
+        }),
+      }),
+    );
+  },
+});
+
+function getPageFormFieldLoaders(appTreeApi: typeof appTreeApiRef.T) {
+  const { tree } = appTreeApi.getTree();
+  const pageNode = tree.nodes.get('page:scaffolder');
+  if (!pageNode?.instance) {
+    return [];
+  }
+  const formFieldNodes = pageNode.edges.attachments.get('formFields') ?? [];
+  return formFieldNodes.flatMap(node => {
+    if (!node.instance) {
+      return [];
+    }
+    const loader = node.instance.getData(
+      FormFieldBlueprint.dataRefs.formFieldLoader,
+    );
+    return loader ? [loader] : [];
+  });
+}
 
 export const scaffolderApi = ApiBlueprint.make({
   params: defineParams =>
@@ -44,4 +103,9 @@ export const scaffolderApi = ApiBlueprint.make({
     }),
 });
 
-export default [scaffolderApi];
+export {
+  formFieldsApiRef,
+  type ScaffolderFormFieldsApi,
+} from '@backstage/plugin-scaffolder-react/alpha';
+
+export default [formFieldsApi, scaffolderApi];
